requirements.txt
flask>=3.0
pillow>=10.0
requests>=2.31
pytest>=8.0
